Update gatling test
Modifications:
- Upgraded Gatling to 2.2.4
- Changed package name from com.excilys to io.resourcepool
- Clean project
