+++ gatling-test/src/test/scala/IDEPathHelper.scala
@@ -1,6 +1,6 @@
 import java.nio.file.Path
 
-import io.gatling.core.util.PathHelper._
+import io.gatling.commons.util.PathHelper._
 
 /**
   * Created by Cédric Cousseran on 29/03/16.
